+++ src/main/java/me/mrCookieSlime/Slimefun/Objects/SlimefunItem/items/Multimeter.java
@@ -6,8 +6,8 @@ import org.bukkit.block.Block;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 
-import io.github.thebusybiscuit.cscorelib2.chat.ChatColors;
 import io.github.thebusybiscuit.cscorelib2.math.DoubleHandler;
+import me.mrCookieSlime.Slimefun.SlimefunPlugin;
 import me.mrCookieSlime.Slimefun.Lists.RecipeType;
 import me.mrCookieSlime.Slimefun.Objects.Category;
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SimpleSlimefunItem;
@@ -28,13 +28,16 @@ public class Multimeter extends SimpleSlimefunItem<ItemUseHandler> {
     		
     		if (block.isPresent()) {
     			Block b = block.get();
+    			
     			if (ChargableBlock.isChargable(b)) {
     				e.cancel();
     				
+    				String stored = DoubleHandler.getFancyDouble(ChargableBlock.getCharge(b)) + " J";
+    				String capacity = DoubleHandler.getFancyDouble(ChargableBlock.getMaxCharge(b)) + " J";
+    				
     				Player p = e.getPlayer();
     				p.sendMessage("");
-    				p.sendMessage(ChatColors.color("&bStored Energy: &3" + DoubleHandler.getFancyDouble(ChargableBlock.getCharge(b)) + " J"));
-    				p.sendMessage(ChatColors.color("&bCapacity: &3" + DoubleHandler.getFancyDouble(ChargableBlock.getMaxCharge(b)) + " J"));
+    				SlimefunPlugin.getLocal().sendMessage(p, "messages.multimeter", str -> str.replace("%stored%", stored).replace("%capacity%", capacity));
     				p.sendMessage("");
     			}
     		}