+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 * Added a config setting to limit how many Nodes a Network can have
 * Added support for Furnaces and Brewing Stands to Cargo Networks
 * Added Organic Food/Fertilizer for Dried Kelp
+* Added many more strings to the messages.yml
 
 ### Changes
 * Removed Solar Array
@@ -54,6 +55,8 @@
 * Fixed #1212
 * Fixed #114
 * Fixed #1385
+* Fixed #1390
+* Fixed #1394
 
 ## Release Candidate 4 (06 Jan 2020)
 https://thebusybiscuit.github.io/builds/TheBusyBiscuit/Slimefun4/stable/#4